requirements.txt
# Scraping and browser automation
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
lxml==4.9.3

# Web framework
streamlit==1.28.0
streamlit-extras==0.3.2

# AI integration
google-generativeai==0.3.1
langchain==0.0.339
langchain-core==0.1.5

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
tqdm==4.66.1

# Image processing
Pillow==10.1.0

# Data manipulation
pandas==2.1.2
numpy==1.26.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
responses==0.24.1

# Security
cryptography==41.0.5
